GS divided to Controller and Connector
