+++ lindeck-frontend/src/views/boody/userPage/UserPage.jsx
@@ -3,17 +3,17 @@ import React from "react";
 import "./UserPage.css"
 import UserCard from "./userCard/UserCard";
 import Paper from "@material-ui/core/Paper";
-import GS from "../../../common/classes/GlobalStorage";
+import Controller from "../../../common/classes/ControllerObject";
 import {Redirect} from "@reach/router";
 
 
 export default class UserPage extends React.Component {
     username = this.props.username
-    user = GS.getUser(this.props.username)
+    user = Controller.getUser(this.props.username)
 
     updateUrl() {
         if (this.username !== this.props.username) {
-            this.user = GS.getUser(this.props.username)
+            this.user = Controller.getUser(this.props.username)
             this.username = this.props.username
         }
     }
@@ -22,23 +22,23 @@ export default class UserPage extends React.Component {
         this.updateUrl()
         if (!this.user) return <Redirect to="/not-found" noThrow/>;
         return <div className="rootUserPage backGroundImage">
-            <UserCard user={this.user} isMe={GS.session.isMe(this.user.id)}/>
+            <UserCard user={this.user} isMe={Controller.session.isMe(this.user.id)}/>
             <Paper className="titlePaper">
-                {GS.session.isOnline && <span>
+                {Controller.session.isOnline && <span>
                 Following : {this.user.following.length}
                 </span>}
-                {!GS.session.isOnline && <span>
+                {!Controller.session.isOnline && <span>
                 Session is offline. Connect to see following.
                 </span>}
             </Paper>
-            {GS.session.isOnline && this.generateFollowing(this.user.following)}
+            {Controller.session.isOnline && this.generateFollowing(this.user.following)}
         </div>
     }
 
     generateFollowing = (followingArray) => {
         return followingArray.map((userNameToFollow, uniqId) => {
-            let user = GS.getUser(userNameToFollow)
-            return <UserCard user={user} key={uniqId} isMe={GS.session.isMe(user.id)}/>
+            let user = Controller.getUser(userNameToFollow)
+            return <UserCard user={user} key={uniqId} isMe={Controller.session.isMe(user.id)}/>
         })
     }
 }
